openchrome: gallium driver for VIA hardware
Start of Gallium driver for VIA hardware. Currently it is a little
skeleton driver but the plan it to expand it to suport glamour with
our xorg driver. This way we don't need to make a specific libdrm
sub library but instead just work on one code base. The plan is make
the driver core code be platform independent and winsys contain the
libdrm type code.

Signed-off-by: James Simmons <jsimmons@infradead.org>